Updates from mainline

+++ src/include/DStoreVec_impl.h
@@ -71,7 +71,7 @@ namespace Loci {
     allocate(e) ;
 
     FORALL(e,ii) {
-      for( int i = 0; i < size; i++) {
+      for( size_t i = 0; i < size; i++) {
         newVec[i] = T() ;
         Loci::streaminput(&newVec[i],1,s) ;
       }
@@ -257,7 +257,7 @@ namespace Loci {
     static_storeVec.allocate(tmp_dom) ;
     static_storeVec.setVecSize(size) ;
     FORALL(tmp_dom, ei) {
-      for(int i = 0; i < size; ++i)
+      for(size_t i = 0; i < size; ++i)
 	static_storeVec[ei][i] = attrib_data[ei][i] ;
     }ENDFORALL ;
     return static_storeVec.Rep() ;
@@ -357,7 +357,7 @@ namespace Loci {
       iter = attrib_data.find(*ci);
       if( iter != attrib_data.end() ) {
         newVec = iter->second;
-        for( int ivec = 0; ivec < size; ivec++){
+        for( size_t ivec = 0; ivec < size; ivec++){
           typename schema_traits::Converter_Type cvtr( newVec[ivec] );
           arraySize += cvtr.getSize();
         }
@@ -420,11 +420,11 @@ namespace Loci {
       iter = attrib_data.find(*ci);
       if( iter == attrib_data.end() ) continue;
       newVec = iter->second;
-      for( int ivec = 0; ivec < size; ivec++){
+      for( size_t ivec = 0; ivec < size; ivec++){
           typename schema_traits::Converter_Type cvtr( newVec[ivec] );
 
           stateSize  = cvtr.getSize();
-          if( stateSize > inbuf.size() ) inbuf.resize(stateSize);
+          if( size_t(stateSize) > inbuf.size() ) inbuf.resize(stateSize);
 
           cvtr.getState( &inbuf[0], stateSize);
           MPI_Pack(&stateSize, 1, MPI_INT, outbuf, outcount,&position,
@@ -448,7 +448,7 @@ namespace Loci {
 
     int M ;
     MPI_Unpack(inbuf, insize, &position, &M, 1, MPI_INT, MPI_COMM_WORLD) ;
-    if( M > size) {
+    if( size_t(M) > size) {
       set_elem_size(M) ;
     }
     unpackdata( traits_type, inbuf, position, insize, seq);
@@ -470,7 +470,7 @@ namespace Loci {
       outcount = size*sizeof(T);
       MPI_Unpack( inbuf, insize, &position, &outbuf[0], outcount, 
                   MPI_BYTE, MPI_COMM_WORLD) ;
-      for( int i = 0; i < size; i++) 
+      for( size_t i = 0; i < size; i++) 
         attrib_data[*ci][i] = outbuf[i];
     }
 
@@ -499,11 +499,11 @@ namespace Loci {
       if( attrib_data[*ci].size() < size) {
         attrib_data[*ci].resize(size);
       }
-      for( int ivec = 0; ivec < size; ivec++) {
+      for( size_t ivec = 0; ivec < size; ivec++) {
         outcount = sizeof(int);
         MPI_Unpack(inbuf, insize, &position, &stateSize, 1, MPI_INT, 
                    MPI_COMM_WORLD) ;
-        if( stateSize > outbuf.size() ) outbuf.resize(stateSize);
+        if( size_t(stateSize) > outbuf.size() ) outbuf.resize(stateSize);
 
         outcount = stateSize*typesize;
         MPI_Unpack(inbuf, insize, &position, &outbuf[0], outcount, MPI_BYTE,
